// packages/modules/blob/src/generated/BlobModule.ts
// Generated by @lynx-js/autolink-codegen. Do not edit.

declare const NativeModules: {
  BlobModule: {
    create(payload: string, type: string): string | null;
    size(blobId: string): number;
    type(blobId: string): string;
    slice(
      blobId: string,
      start: number,
      end: number,
      contentType: string,
    ): string | null;
    bytes(blobId: string): string | null;
    text(blobId: string): string | null;
    arrayBuffer(blobId: string): string | null;
  };
};

export const BlobModule: typeof NativeModules.BlobModule =
  NativeModules.BlobModule;
export default BlobModule;